revert cursor filter changes, keep operand support only
