+++ packages/twenty-front/src/modules/object-record/record-show/hooks/useRecordShowPagePagination.ts
@@ -84,7 +84,6 @@ export const useRecordShowPagePagination = (
         orderBy,
         cursorRecordValues: currentRecordKeysetValues,
         isForwardPagination: false,
-        fieldMetadataItems: objectMetadataItem.fields,
       })
     : undefined;
 
@@ -93,7 +92,6 @@ export const useRecordShowPagePagination = (
         orderBy,
         cursorRecordValues: currentRecordKeysetValues,
         isForwardPagination: true,
-        fieldMetadataItems: objectMetadataItem.fields,
       })
     : undefined;
